Update quiz.py
Line 57 and 146.

Line 57 ensures that the list of each object only belongs to that instance, and isn't shared among multiple different MCQuestions (leading to answers from one question appearing in the other).

146 is changed back to the original. "Wrong(data)" by itself didn't do anything, as that object wasn't saved anywhere. The line currently knows how to handle the data and add it to the MCQuestion.

I made some further changes to line splitting, but that isn't important. I will submit this file as is